Fikset slik at bilder hopper til neste linje hvis det ikke er plass på mobil.
Skilt ut kartkomponenten og gjort det mulig å få opp popup med interaktivt kart også på detaljsida
