+++ src/app/components/observation/observation/observation.component.css
@@ -65,8 +65,8 @@ swiper-container swiper-slide:first-child:nth-last-child(1) {
   width: 100%;
 }
 
-app-static-map-image {
-  width: 100%;
+observation-location-map {
+  height: 256px;
 }
 
 swiper-slide img {
